feat(folderStructure): create generatedFolder to receive generated files
