+++ packages/core/manifest/e2e/assets/mock-manifest.yml
