+++ packages/core/manifest/e2e/jest.mysql-setup.ts
@@ -61,7 +61,7 @@ beforeAll(async () => {
       load: () =>
         load(
           fs.readFileSync(
-            `${process.cwd()}/e2e/assets/mock-backend.yml`,
+            `${process.cwd()}/e2e/assets/mock-manifest.yml`,
             'utf8'
           )
         )
