--- packages/core/manifest/manifest/backend.yml
@@ -1,60 +1,0 @@
-name: Personal portfolio
-entities:
-  User:
-    properties:
-      - name
-    authenticable: true
-    policies:
-      update:
-        - access: public
-      read:
-        - access: public
-
-  Project:
-    properties:
-      - { name: title, validation: { required: true } }
-      - { name: excerpt, validation: { required: true } }
-      - role
-      - { name: date, type: date }
-      - { name: url, type: link }
-      - {
-          name: photo,
-          type: image,
-          options:
-            {
-              sizes:
-                {
-                  small: { height: 403, width: 805 },
-                  social: { height: 630, width: 1200 },
-                  large: { height: 806, width: 1610 }
-                }
-            }
-        }
-      - { name: description, type: richText }
-      - seoTitle
-      - seoDescription
-    policies:
-      read:
-        - { access: restricted, allow: [User] }
-    belongsTo:
-      - User
-    belongsToMany:
-      - Contact
-
-  Contact:
-    properties:
-      - { name: name, type: string, validation: { required: true } }
-      - { name: email, type: email, validation: { required: true } }
-      - { name: message, type: text, validation: { required: true } }
-    policies:
-      create:
-        - access: public
-
-  Dog:
-    properties:
-      - { name: name, type: string, validation: { required: true } }
-      - { name: age, type: number, validation: { required: true } }
-      - { name: photo, type: image }
-
-    belongsToMany:
-      - User
