+++ packages/create-manifest/assets/manifest.yml
@@ -1,4 +1,4 @@
-# This is a sample file for a backend.yml manifest backend.
+# This is a sample file for a manifest.yml manifest backend.
 # Read more about the manifest format here: https:/manifest.build/docs
 
 name: My pet app 🐾
